+++ src/components/SendAsset/SendAsset.js
@@ -25,29 +25,31 @@ import { createStructuredSelector } from 'reselect';
 import t from 'translations/translate';
 import { BigNumber } from 'bignumber.js';
 
-// actions
+// Actions
 import { estimateTransactionAction, resetEstimateTransactionAction } from 'actions/transactionEstimateActions';
 
-// constants
+// Constants
 import { SEND_COLLECTIBLE_CONFIRM, SEND_TOKEN_CONFIRM } from 'constants/navigationConstants';
 import { ASSET_TYPES } from 'constants/assetsConstants';
+import { REMOTE_CONFIG } from 'constants/remoteConfigConstants';
 
-// components
+// Components
 import Button from 'components/legacy/Button';
 import FeeLabelToggle from 'components/FeeLabelToggle';
 import Toast from 'components/Toast';
 
-// utils
+// Utils
 import { truncateAmount, reportErrorLog } from 'utils/common';
 import { getBalanceBN, isEnoughBalanceForTransactionFee } from 'utils/assets';
 import { isValidValueForTransfer } from 'utils/transactions';
+import { openUrl } from 'utils/inAppBrowser';
 
-// selectors
+// Selectors
 import { useGasTokenSelector } from 'selectors/archanova';
 import { contactsSelector } from 'selectors';
 import { accountAssetsWithBalanceSelector } from 'selectors/assets';
 
-// types
+// Types
 import type { NavigationScreenProp } from 'react-navigation';
 import type {
   TransactionPayload,
@@ -61,6 +63,9 @@ import type { SessionData } from 'models/Session';
 import type { Contact } from 'models/Contact';
 import type { AccountAssetBalances } from 'models/Balances';
 import type { Chain } from 'models/Chain';
+
+// Services
+import { firebaseRemoteConfig } from 'services/firebase';
 
 // Local
 import SendContainer from './SendContainer';
@@ -106,6 +111,8 @@ const SendAsset = ({
   const [value, setValue] = useState(null);
   const [selectedContact, setSelectedContact] = useState(defaultContact);
   const [submitPressed, setSubmitPressed] = useState(false);
+
+  const transactionRevertedAricleURL = firebaseRemoteConfig.getString(REMOTE_CONFIG.TRANSACTION_REVERTED_ARTICLE_URL);
 
   const assetAddress = assetData.contractAddress;
   const chain = assetData?.chain;
@@ -156,9 +163,12 @@ const SendAsset = ({
     if (!feeInfo || !selectedContact || !assetData) {
       // something went wrong
       Toast.show({
-        message: t('toast.cannotSendAsset'),
-        emoji: 'woman-shrugging',
-        supportLink: true,
+        message: t('toast.transactionReverted'),
+        emoji: 'hushed',
+        autoClose: false,
+        onPress: () => {
+          openUrl(transactionRevertedAricleURL);
+        },
       });
       reportErrorLog('SendAsset screen handleFormSubmit failed', { feeInfo, selectedContact, assetData });
       return;
