+++ src/constants/remoteConfigConstants.js
@@ -51,6 +51,7 @@ export const REMOTE_CONFIG = {
   PRISMIC_TERMS_OF_POLICY_DOCUMENT_ID: 'prismic_terms_of_policy_document_id',
   PRISMIC_INTERJECTION_DOCUMENT_ID: 'prismic_interjection_document_id',
   FEATURE_WC_DASHBOARD_INAPPBROWSER: 'feature_wc_dashboard_inappbrowser',
+  TRANSACTION_REVERTED_ARTICLE_URL: 'transaction_reverted_article_url',
 };
 
 // These are used as a fallback in case firebase fails to fetch actual values
@@ -84,4 +85,5 @@ export const INITIAL_REMOTE_CONFIG = {
   [REMOTE_CONFIG.PRISMIC_TERMS_OF_POLICY_DOCUMENT_ID]: 'YNGo8hIAACMAsjRR',
   [REMOTE_CONFIG.PRISMIC_INTERJECTION_DOCUMENT_ID]: 'YNsHjxIAACIAw0ca',
   [REMOTE_CONFIG.FEATURE_WC_DASHBOARD_INAPPBROWSER]: false,
+  [REMOTE_CONFIG.TRANSACTION_REVERTED_ARTICLE_URL]: 'https://help.pillar.fi/en/articles/5556539-why-transactions-get-reverted',
 };
