+++ src/screens/Exchange/ExchangeConfirm/ExchangeConfirm.js
@@ -34,6 +34,7 @@ import TransactionDeploymentWarning from 'components/other/TransactionDeployment
 // Constants
 import { SEND_TOKEN_PIN_CONFIRM } from 'constants/navigationConstants';
 import { TRANSACTION_TYPE } from 'constants/transactionsConstants';
+import { REMOTE_CONFIG } from 'constants/remoteConfigConstants';
 
 // Hooks
 import { useTransactionsEstimate, useTransactionFeeCheck } from 'hooks/transactions';
@@ -50,9 +51,13 @@ import { getFormattedBalanceInFiat } from 'utils/assets';
 import { useProviderConfig } from 'utils/exchange';
 import { mapTransactionsToTransactionPayload } from 'utils/transactions';
 import { spacing } from 'utils/variables';
+import { openUrl } from 'utils/inAppBrowser';
 
 // Types
 import type { ExchangeOffer } from 'models/Exchange';
+
+// Services
+import { firebaseRemoteConfig } from 'services/firebase';
 
 // Local
 import ExchangeScheme from './ExchangeScheme';
@@ -72,6 +77,8 @@ const ExchangeConfirmScreen = () => {
 
   const chainRates = useChainRates(chain);
 
+  const transactionRevertedAricleURL = firebaseRemoteConfig.getString(REMOTE_CONFIG.TRANSACTION_REVERTED_ARTICLE_URL);
+
   const { feeInfo, errorMessage: estimationErrorMessage, isEstimating } = useTransactionsEstimate(
     chain,
     offer.transactions,
@@ -83,9 +90,12 @@ const ExchangeConfirmScreen = () => {
   const confirmTransaction = () => {
     if (!feeInfo) {
       Toast.show({
-        message: t('toast.cannotExchangeAsset'),
-        emoji: 'woman-shrugging',
-        supportLink: true,
+        message: t('toast.transactionReverted'),
+        emoji: 'hushed',
+        autoClose: false,
+        onPress: () => {
+          openUrl(transactionRevertedAricleURL);
+        },
       });
 
       return;
